repalced old home page with beta page and removed beta
