+++ src/pages/index.js
@@ -2,58 +2,69 @@ import React from 'react'
 import Helmet from 'react-helmet'
 import { connect } from 'react-redux'
 import Link from '../components/utils/temp-link'
-
+import { withPrefix as withPrefixGatsby } from 'gatsby-link'
+import { hydrate as hydateDataManagerAction } from '../state/reducers/data-sets'
+import { normalize as normalizeDataSetAction } from '../state/reducers/data-sets'
 import {
-  hydrate as hydateDataManagerAction,
-  normalize as normalizeDataSetAction,
   PRODUCT_VOLUMES_FISCAL_YEAR,
   REVENUES_MONTHLY,
   REVENUES_FISCAL_YEAR,
+  BY_ID, BY_COMMODITY,
+  BY_STATE, BY_COUNTY,
+  BY_OFFSHORE_REGION,
+  BY_LAND_CATEGORY,
+  BY_LAND_CLASS,
+  BY_REVENUE_TYPE,
   BY_FISCAL_YEAR,
   BY_CALENDAR_YEAR
 } from '../state/reducers/data-sets'
 
 import * as CONSTANTS from '../js/constants'
-import { KeyStatsSection } from '../components/sections/KeyStatsSection'
-import { MapSection } from '../components/sections/MapSection'
-import { WhatsNew } from '../components/sections/WhatsNew'
-import { Tabordion, Tab } from '../components/layouts/Tabordion'
-import StateMap from '../components/maps/StateMap'
-import FederalLandOwnershipLegend from '../components/maps/FederalLandOwnershipLegend'
-import LocationSelector from '../components/selectors/LocationSelector'
-import { ExploreDataButton } from '../components/layouts/buttons/ExploreDataButton'
-import { BlueButton } from '../components/layouts/buttons/BlueButton'
-import { ExploreDataLink } from '../components/layouts/icon-links/ExploreDataLink'
-import { DownloadDataLink } from '../components/layouts/icon-links/DownloadDataLink'
-import MapLink from '../components/layouts/icon-links/MapLink'
-import RevenueTrends from '../components/sections/RevenueTrends'
-import DisbursementTrends from '../components/sections/DisbursmentTrends'
+import utils from '../js/utils'
+import GlossaryTerm from '../components/utils/glossary-term.js';
+
 
 import DefaultLayout from '../components/layouts/DefaultLayout'
+import { Tabordion, Tab } from '../components/layouts/Tabordion'
+import TabContainer from '../components/layouts/Tabordion/TabContainer.js'
+import RevenueTrends from '../components/sections/RevenueTrends'
+import TotalRevenue from '../components/sections/TotalRevenue/TotalRevenueDeprecated'
+import TotalDisbursements from '../components/sections/TotalDisbursements/TotalDisbursementsDeprecated'
+import TotalProduction from '../components/sections/TotalProduction/TotalProductionDeprecated'
+import DisbursementTrends from '../components/sections/DisbursmentTrends'
+import ExploreRevenue from '../components/sections/Explore/Revenue'
+import ExploreDisbursements from '../components/sections/Explore/Disbursements'
+import ExploreProduction from '../components/sections/Explore/Production'
+import { MapSection } from '../components/sections/MapSection'
 
-import styles from './index.module.css'
+import { WhatsNew } from '../components/sections/WhatsNew'
+let mapJson = require("../../static/maps/land/us-topology.json")
+let  mapOffshoreJson =require("../../static/maps/offshore/offshore.json")
 
-class HomePage extends React.Component {
+
+class Beta extends React.Component {
   constructor (props) {
     super(props)
 
     this.hydrateStore()
   }
 
+
+    
   /**
    * Add the data to the redux store to enable
    * the components to access filtered data using the
    * reducers
    **/
   hydrateStore () {
-    let data = this.props.data
+    let data = this.props.data;
 
     this.props.normalizeDataSet([
       { key: REVENUES_MONTHLY,
         data: data.allMonthlyRevenues.data,
         groups: [
           {
-            key: BY_CALENDAR_YEAR,
+            key:BY_CALENDAR_YEAR,
             groups: data.allMonthlyRevenuesByCalendarYear.group
           }
         ]
@@ -62,7 +73,7 @@ class HomePage extends React.Component {
         data: data.allFiscalYearRevenues.data,
         groups: [
           {
-            key: BY_FISCAL_YEAR,
+            key:BY_FISCAL_YEAR,
             groups: data.allFiscalYearRevenuesByFiscalYear.group
           }
         ]
@@ -72,20 +83,20 @@ class HomePage extends React.Component {
         data: data.allFiscalYearProductVolumes.data,
         groups: [
           {
-            key: BY_FISCAL_YEAR + '_Gas',
+            key:BY_FISCAL_YEAR+"_Gas",
             groups: data.allFiscalYearProductVolumesByFiscalYear_Gas.group
           },
           {
-            key: BY_FISCAL_YEAR + '_Oil',
+            key:BY_FISCAL_YEAR+"_Oil",
             groups: data.allFiscalYearProductVolumesByFiscalYear_Oil.group
           },
           {
-            key: BY_FISCAL_YEAR + '_Coal',
+            key:BY_FISCAL_YEAR+"_Coal",
             groups: data.allFiscalYearProductVolumesByFiscalYear_Coal.group
           }
         ]
       }
-    ])
+    ]);
 
     this.props.hydateDataManager([
       { key: CONSTANTS.PRODUCTION_VOLUMES_OIL_KEY, data: this.props.data.OilVolumes.volumes },
@@ -96,134 +107,93 @@ class HomePage extends React.Component {
     ])
   }
 
+//const Beta = (props) => {
     render () {
-    return (
-      <DefaultLayout>
-        <main id="main-content">
-          <Helmet
-            title="Home | Natural Resources Revenue Data"
-            meta={[
-              // title
-              { name: 'og:title', content: 'Home | Natural Resources Revenue Data' },
-              { name: 'twitter:title', content: 'Home | Natural Resources Revenue Data' },
-            ]}
-
-          />
-          <Tabordion>
-            <Tab id="tab-overview" name="Overview">
-              <div className={styles.tabContentContainer} >
-                <div className={styles.tabContent}>
-                  <p>When companies extract natural resources on federal lands and offshore areas, they pay bonuses, rent, and royalties to the federal government. The Office of Natural Resources Revenue (ONRR) distributes these funds for public use in a variety of ways.</p>
-                  <div className={styles.tabContentBottomContainer}>
-                    <div>
-                      <ExploreDataButton />
-                    </div>
-                    <div>
-                      <Link to="/how-it-works">Learn how it works</Link>
-                    </div>
-                    <div className={styles.tabContentBottomContainerLinks}>
-                      <div>
-                        <ExploreDataLink to="/explore/revenue">Revenue data</ExploreDataLink>
-                      </div>
-                      <div>
-                        <MapLink to="#map-section" />
-                      </div>
-                    </div>
-                  </div>
-                </div>
-                <div className={styles.tabContentAside}>
-                  <RevenueTrends />
-                </div>
-              </div>
-            </Tab>
-            <Tab id="tab-production" name="Production">
-              <div className={styles.tabContentContainer} >
-                <div className={styles.tabContent} >
-                  <p>The United States is among the world's top producers of oil, natural gas, and coal. The U.S. is also a global leader in renewable energy production. We have data for energy and mineral production on federal lands and waters and Native American lands.</p>
-                  <div className={styles.tabContentBottomContainer}>
-                    <div>
-                      <BlueButton to="/how-it-works/#production">How production works</BlueButton>
-                    </div>
-                  </div>
-                </div>
-                <div className={styles.tabContentAside}>
-                  <h5>Explore production data</h5>
-                  <div className={styles.linkContainer}>
-                    <ExploreDataLink to="/explore/#federal-production">Federal lands and waters</ExploreDataLink>
-                    <ExploreDataLink to="/how-it-works/native-american-production/#production-on-native-american-land">Native American lands</ExploreDataLink>
-                    <DownloadDataLink to="/downloads/#production">Downloads and documentation</DownloadDataLink>
-                  </div>
-                </div>
-              </div>
-            </Tab>
-            <Tab id="tab-revenue" name="Revenue">
-              <div className={styles.tabContentContainer} >
-                <div className={styles.tabContent} >
-                  <p>Companies pay to extract natural resources on federal lands and waters. These payments include bonuses, rents, and royalties, resulting in revenue to the American public. The Office of Natural Resources Revenue (ONRR) collects and distributes this revenue.</p>
-                  <div className={styles.tabContentBottomContainer}>
-                    <div>
-                      <BlueButton to="/how-it-works/revenues">How revenue works</BlueButton>
-                    </div>
-                  </div>
-                </div>
-                <div className={styles.tabContentAside}>
-                  <h5>Explore revenue data</h5>
-                  <div className={styles.linkContainer}>
-                    <ExploreDataLink to="/explore/revenue">Revenue in detail</ExploreDataLink>
-                    <ExploreDataLink to="/how-it-works/federal-revenue-by-company/">Revenue by company</ExploreDataLink>
-                    <DownloadDataLink to="/downloads/#revenue">Downloads and documentation</DownloadDataLink>
-                  </div>
-                </div>
-              </div>
-            </Tab>
-            <Tab id="tab-disbursements" name="Disbursements">
-              <div className={styles.tabContentContainer} >
-                <div className={styles.tabContent} >
-                  <p>After collecting revenue from natural resource extraction, the Office of Natural Resources Revenue (ONRR) distributes that money to different agencies, funds, and local governments for public use. This process is called “disbursement.”</p>
-                  <div className={styles.tabContentBottomContainer}>
-                    <div>
-                      <BlueButton to="/how-it-works/#disbursements">How disbursements work</BlueButton>
-		      <div className={styles.tabContentBottomContainerLinks}>
-			<h5>Explore disbursements data</h5>
-			<div className={styles.linkContainer}>
-			  <ExploreDataLink to="/explore/#by-fund">By recipient</ExploreDataLink>
-			  <DownloadDataLink to="/downloads/disbursements">Downloads and documentation</DownloadDataLink>
-			</div>
-		      </div>
-                    </div>
-                  </div>
-                </div>
-                <div className={styles.tabContentAside}>
-                  <DisbursementTrends />
-                </div>
-              </div>
-            </Tab>
-          </Tabordion>
-
-          <KeyStatsSection />
-          <MapSection b
-               title="Learn about extractive industries in each state"
-               info="Explore production, revenue, and disbursements data for each state."
-               states={this.props.data.states_data.states}
-               offshore_regions={this.props.data.offshore_data.offshore_regions}
-               mapFeatures="states"
-               mapTitle="Revenue"
-               mapJson="/maps/land/us-topology.json"
-               mapOffshoreJson="/maps/offshore/offshore.json"
-               onClick={ (d,i) => { 
-                   let state=fipsAbbrev[d.id] || d.id;
-                   let url="/explore/"+state+"#revenue" 
-                   window.location.href = url;
-
- } }
-/>
-
-
-          <WhatsNew />
-        </main>
-      </DefaultLayout>
-    )
-  }
+	console.debug(mapJson);
+	return(
+	    <DefaultLayout>
+	      <main>
+		<Helmet
+		  title="Beta | Natural Resources Revenue Data"
+		  meta={[
+		      // title
+		      { name: 'og:title', content: 'Beta | Natural Resources Revenue Data' },
+		      { name: 'twitter:title', content: 'Beta | Natural Resources Revenue Data' },
+		  ]} >
+		</Helmet>
+	        <section className="container-page-wrapper" >
+		<h3 className="h3-bar"></h3>
+		<p> When companies extract energy and mineral resources on property leased from the federal government and Native Americans, they pay <GlossaryTerm termKey="Bonus">bonuses</GlossaryTerm>, <GlossaryTerm>rent</GlossaryTerm>, and <GlossaryTerm termKey="Royalty">royalties</GlossaryTerm>. The Office of Natural Resources Revenue (ONRR) collects and <GlossaryTerm termKey="disbursement">disburses</GlossaryTerm> revenue from federal lands and waters to different agencies, funds, and local governments for public use. All revenue collected from extraction on Native American lands is disbursed to Native American tribes, nations, or individuals.</p>
+		</section>
+		<Tabordion>
+		<Tab id="tab-revenue" name="Revenue">
+		    <TabContainer id="tab-container-revenue" name="Revenue"
+				  title="Revenue"
+				  info="The amount of money collected by the federal government from energy and mineral extraction on federal lands and waters and Native American lands."
+				  contentLeft={<TotalRevenue/>}
+				  contentRight={<RevenueTrends/>}
+				  contentBottom={<ExploreRevenue />}
+				  />
+		  </Tab>
+		  <Tab id="tab-disbursements" name="Disbursements" >
+		    <TabContainer id="tab-container-disbursements" 
+				  title="Disbursements"
+				  info="The amount of money the federal government distributed to various funds, agencies, local governments, and Native Americans."
+				  contentRight={<DisbursementTrends/>}
+         			  contentLeft={<TotalDisbursements/>}
+				  contentBottom={<ExploreDisbursements/>}
+				  
+				  />
+		    
+		  </Tab>  
+		  <Tab id="tab-production" name="Production" >
+		    <TabContainer id="tab-container-production" 
+			      title="Production"
+				  info="The volume of major commodities extracted on federal lands and waters and Native American lands."
+				  contentBottom={<ExploreProduction/>}
+				  >
+				  <TotalProduction />
+		    </TabContainer>
+		    
+		    
+		  </Tab>  
+		  <Tab id="tab-by-state" name="Data by state" >
+		    <TabContainer id="tab-container-by-state" 
+				  title="Data by state"
+				  
+				  >
+      <p><em>Select a state for detailed production, revenue, and disbursements data.</em></p>
+		      <MapSection 
+			info="Federal revenue by state and offshore region for fiscal year 2018"
+			states={this.props.data.states_data.states}
+			offshore_regions={this.props.data.offshore_data.offshore_regions}
+			mapFeatures="states"
+			mapTitle="Revenue"
+			mapJson="/maps/land/us-topology.json"
+			mapOffshoreJson="/maps/offshore/offshore.json"
+			mapJsonObject={{us:mapJson, offshore:mapOffshoreJson}}
+			onClick={ (d,i) => {
+			    let state=fipsAbbrev[d.id] || d.id;
+			    let url="/explore/"+state+"#revenue" 
+			    if(state.match(/offshore/)) {
+				url="/explore/"+state;
+			    }
+			    url=withPrefixGatsby(url);
+			    window.location.href = url;
+			    
+			} }
+			/>
+		    </TabContainer>
+		  </Tab>  
+		</Tabordion>	
+		
+		
+		
+		<WhatsNew />
+	      </main>
+	    </DefaultLayout>
+	)
+    }
 }
 
 export default connect(
@@ -232,11 +202,28 @@ export default connect(
     hydateDataManager: dataSets => dispatch(hydateDataManagerAction(dataSets)),
     normalizeDataSet: dataSets => dispatch(normalizeDataSetAction(dataSets))
   })
-)(HomePage)
+)(Beta)
 
-export const query = graphql `
-  query HomePageQuery {
-    offshore_data:allMarkdownRemark (filter:{fileAbsolutePath: {regex: "/offshore_regions/"}} sort:{fields: [frontmatter___title], order: ASC}) {
+/*
+		<TabContainer id="tab-container-revenue" name="Revenue"
+			      title="Revenue"
+			      info="How much money do energy and mineral resources bring into federal government?"
+			      tabContentRight={<RevenueTrends/>}
+         		      tabContentLeft={<RevenueTrends/>}
+			      />
+		<TabContainer id="tab-container-disbursements" 
+		  title="Disbursements"
+		  info="Where does the money that is brought in go?"
+		  tabContentRight={<DisbursementTrends/>}
+         	  tabContentLeft={<DisbursementTrends/>}
+		  
+		  />
+*/
+
+
+export const query = graphql`
+  query BetaQuery {
+   offshore_data:allMarkdownRemark (filter:{fileAbsolutePath: {regex: "/offshore_regions/"}} sort:{fields: [frontmatter___title], order: ASC}) {
       offshore_regions:edges {
         offshore_region:node {
           frontmatter {
@@ -468,6 +455,7 @@ export const query = graphql `
   }
 `
 
+
 const fipsAbbrev={
  "02":  "AK",
  "01":  "AL",
@@ -524,4 +512,34 @@ const fipsAbbrev={
  "55":  "WI",
  "54":  "WV",
  "56":  "WY",
+    "ALA":"offshore-alaska",
+    "ALB":"offshore-alaska",
+    "BFT":"offshore-alaska",
+    "BOW":"offshore-alaska",
+    "CHU":"offshore-alaska",
+    "COK":"offshore-alaska",
+    "GEO":"offshore-alaska",
+    "GOA":"offshore-alaska",
+    "HOP":"offshore-alaska",
+    "KOD":"offshore-alaska",
+    "MAT":"offshore-alaska",
+    "NAL":"offshore-alaska",
+    "NAV":"offshore-alaska",
+    "NOR":"offshore-alaska",
+    "SHU":"offshore-alaska",
+
+    "FLS":"offshore-atlantic",
+    "MDA":"offshore-atlantic",
+    "NOA":"offshore-atlantic",
+    "SOA":"offshore-atlantic",
+
+    "WGM":"offshore-gulf",
+    "CGM":"offshore-gulf",
+    "EGM":"offshore-gulf",
+
+    "CEC":"offshore-pacific",
+    "NOC":"offshore-pacific", 
+    "SOC":"offshore-pacific",
+    "WAO":"offshore-pacific"
+
 }
